+++ modules/oit/include/modules/oit/processors/linerasterizer.h
@@ -70,7 +70,7 @@ public:
     virtual const ProcessorInfo& getProcessorInfo() const override;
     static const ProcessorInfo processorInfo_;
 
-    virtual void rasterize(const ivec2& imageSize, const mat4& worldMatrixTransform) override;
+    virtual void rasterize(const ivec2& imageSize, const dmat4& worldMatrixTransform) override;
     virtual UseFragmentList usesFragmentLists() const override;
 
     virtual std::optional<dmat4> boundingBox() const override;
